requirements.txt
flask
tornado
pymongo
flask-pymongo
pdfminer
progressbar
itsdangerous
sockjs-tornado>=1.0.2
redis>=2.10.5